Update prefetch mechanism:
Prefetch requests and responses are stored in a ringbuffer instead of
a queue, which means we can utilize prefetches of many relations
concurrently -- page reads of un-prefetched relations now don't
imply dropping prefetches.

In a future iteration, this will detect sequential scans based on the
read behavior of those scans, and will add prefetches for that relation
as needed.